refactor(guard): split native runtime policy modules
