+++ rust/crates/guard-rule-contract/src/lib.rs
@@ -6,7 +6,7 @@ use sha2::{Digest, Sha256};
 pub const RULE_CONTRACT_SCHEMA: &str = "hol-guard-native-rule-contract.v2";
 const RULE_CONTRACT_DOMAIN: &[u8] = b"hol-guard-native-rule-contract.v2\0";
 
-const COMPONENTS: [(&str, &[u8]); 15] = [
+const COMPONENTS: [(&str, &[u8]); 16] = [
     (
         "guard-rules",
         include_bytes!("../../guard-rules/src/lib.rs"),
@@ -50,6 +50,10 @@ const COMPONENTS: [(&str, &[u8]); 15] = [
     (
         "guard-runtime-policy-enforcement-facts",
         include_bytes!("../../guard-runtime/src/policy_enforcement_facts.rs"),
+    ),
+    (
+        "guard-runtime-policy-enforcement-facts-tools",
+        include_bytes!("../../guard-runtime/src/policy_enforcement_facts_tools.rs"),
     ),
     (
         "guard-runtime-policy-enforcement-policy",
@@ -143,6 +147,7 @@ mod tests {
                 "guard-command-pretool-extract",
                 "guard-runtime-policy-enforcement",
                 "guard-runtime-policy-enforcement-facts",
+                "guard-runtime-policy-enforcement-facts-tools",
                 "guard-runtime-policy-enforcement-policy",
                 "guard-policy-snapshot",
                 "guard-policy-snapshot-canonical",
