+++ rust/crates/guard-runtime/src/main.rs
@@ -7,28 +7,28 @@ mod oneshot;
 mod policy_enforcement;
 mod policy_store;
 mod resident_client;
+mod resident_protocol;
 mod resident_state;
 mod resident_state_encoding;
+mod resident_transport;
+mod resident_transport_service;
 mod strict_json;
 
-use guard_command::CommandModelRequestV1;
-use guard_contracts::{
-    GuardHookEnvelopeV2, NativeHookRequestV1, RuntimeCapabilitiesV1, GUARD_HOOK_ENVELOPE_V2_SCHEMA,
-    MAX_NATIVE_REQUEST_BYTES, MAX_NATIVE_RESPONSE_BYTES, NATIVE_PROTOCOL_VERSION,
+pub(crate) use resident_protocol::{
+    capabilities, encode_response, error_response, evaluate_resident_bytes, strict_json_value,
 };
-use guard_hook_core::review_post_tool;
-use policy_store::PolicySnapshotStore;
-use serde::Deserialize;
-use serde_json::Value;
-use sha2::{Digest, Sha256};
+pub(crate) use resident_transport::{
+    admit_connection, constant_time_eq, hmac_sha256, start_resident_workers, BoxedResidentStream,
+    ResidentStream,
+};
+pub(crate) use resident_transport_service::{
+    read_resident_auth_token, resident_stdin_liveness, serve, serve_loopback,
+};
+
+use guard_contracts::MAX_NATIVE_REQUEST_BYTES;
 use std::env;
-use std::io::{self, BufRead, Read, Write};
-use std::net::{Ipv4Addr, SocketAddr, TcpListener, TcpStream};
-use std::panic::{catch_unwind, AssertUnwindSafe};
-use std::sync::mpsc::{sync_channel, Receiver, SyncSender, TrySendError};
-use std::sync::{Arc, Mutex};
-use std::thread;
-use std::time::{Duration, Instant};
+use std::io::{self, Read, Write};
+use std::time::Duration;
 
 const BUILD_SHA: &str = match option_env!("HOL_GUARD_BUILD_SHA") {
     Some(value) => value,
@@ -61,129 +61,6 @@ const CLIENT_PROOF_LABEL: &[u8] = b"hol-guard-resident-client-v1\0";
 #[cfg(unix)]
 const PARENT_LIVENESS_FD_ENV: &str = "HOL_GUARD_PARENT_LIVENESS_FD";
 
-#[derive(Debug, Deserialize)]
-#[serde(tag = "operation", content = "request", rename_all = "snake_case")]
-enum ResidentOperationV1 {
-    CommandModel(CommandModelRequestV1),
-    PreToolUse(CommandModelRequestV1),
-    PolicySnapshotPush(Value),
-    Health(Value),
-    Shutdown(Value),
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(untagged)]
-enum ResidentRequestV1 {
-    Operation(ResidentOperationV1),
-    Edge(GuardHookEnvelopeV2),
-    Hook(NativeHookRequestV1),
-}
-
-trait ResidentStream: Read + Write + Send {
-    fn set_resident_read_timeout(&self, timeout: Option<Duration>) -> io::Result<()>;
-    fn set_resident_write_timeout(&self, timeout: Option<Duration>) -> io::Result<()>;
-    fn try_read_available(&mut self, output: &mut [u8]) -> io::Result<usize>;
-    fn configure_low_latency(&self) -> io::Result<()> {
-        Ok(())
-    }
-}
-
-impl ResidentStream for TcpStream {
-    fn set_resident_read_timeout(&self, timeout: Option<Duration>) -> io::Result<()> {
-        TcpStream::set_read_timeout(self, timeout)
-    }
-
-    fn set_resident_write_timeout(&self, timeout: Option<Duration>) -> io::Result<()> {
-        TcpStream::set_write_timeout(self, timeout)
-    }
-
-    fn try_read_available(&mut self, output: &mut [u8]) -> io::Result<usize> {
-        self.set_nonblocking(true)?;
-        let result = self.read(output);
-        let restore = self.set_nonblocking(false);
-        match (result, restore) {
-            (Err(error), Ok(())) if error.kind() == io::ErrorKind::WouldBlock => Ok(0),
-            (result, Ok(())) => result,
-            (_, Err(error)) => Err(error),
-        }
-    }
-
-    fn configure_low_latency(&self) -> io::Result<()> {
-        self.set_nodelay(true)
-    }
-}
-
-#[cfg(unix)]
-impl ResidentStream for std::os::unix::net::UnixStream {
-    fn set_resident_read_timeout(&self, timeout: Option<Duration>) -> io::Result<()> {
-        std::os::unix::net::UnixStream::set_read_timeout(self, timeout)
-    }
-
-    fn set_resident_write_timeout(&self, timeout: Option<Duration>) -> io::Result<()> {
-        std::os::unix::net::UnixStream::set_write_timeout(self, timeout)
-    }
-
-    fn try_read_available(&mut self, output: &mut [u8]) -> io::Result<usize> {
-        self.set_nonblocking(true)?;
-        let result = self.read(output);
-        let restore = self.set_nonblocking(false);
-        match (result, restore) {
-            (Err(error), Ok(())) if error.kind() == io::ErrorKind::WouldBlock => Ok(0),
-            (result, Ok(())) => result,
-            (_, Err(error)) => Err(error),
-        }
-    }
-}
-
-type BoxedResidentStream = Box<dyn ResidentStream>;
-
-struct PendingRequest {
-    stream: BoxedResidentStream,
-    request_id: [u8; FRAME_REQUEST_ID_BYTES],
-    request_digest: [u8; FRAME_DIGEST_BYTES],
-    length: usize,
-    payload_prefix: Vec<u8>,
-    accepted_at: Instant,
-}
-
-fn capabilities() -> RuntimeCapabilitiesV1 {
-    let mut features = vec![
-        "post-tool-inline-v1".into(),
-        "post-tool-source-read-v1".into(),
-        "oneshot-v1".into(),
-        "framed-serve-v1".into(),
-        "resident-protocol-v2".into(),
-        "bounded-admission-v2".into(),
-        "overload-signal-v1".into(),
-        "panic-containment-v1".into(),
-        "rule-contract-v2".into(),
-        "pre-tool-command-model-shadow-v1".into(),
-        "resident-command-model-shadow-v1".into(),
-        "pre-tool-command-authority-v1".into(),
-        "pre-tool-generic-authority-v1".into(),
-        "policy-snapshot-v3".into(),
-        "policy-snapshot-push-v1".into(),
-        "native-policy-in-memory-v1".into(),
-        "hook-envelope-v2".into(),
-        "native-resident-client-v1".into(),
-        "native-resident-lifecycle-v1".into(),
-    ];
-    if cfg!(windows) {
-        features.push("authenticated-loopback-resident-v1".into());
-    }
-    if cfg!(unix) {
-        features.push("authenticated-unix-resident-v1".into());
-    }
-    RuntimeCapabilitiesV1 {
-        protocol_version: NATIVE_PROTOCOL_VERSION,
-        runtime_version: PACKAGE_VERSION.to_owned(),
-        rule_digest: guard_rule_contract::rule_digest(),
-        build_sha: BUILD_SHA.to_owned(),
-        target: format!("{}-{}", env::consts::ARCH, env::consts::OS),
-        features,
-    }
-}
-
 fn read_stdin_bounded() -> Result<Vec<u8>, String> {
     let mut bytes = Vec::new();
     io::stdin()
@@ -196,537 +73,11 @@ fn read_stdin_bounded() -> Result<Vec<u8>, String> {
     Ok(bytes)
 }
 
-pub(crate) fn strict_json_value(bytes: &[u8]) -> Result<Value, String> {
-    strict_json::parse(bytes)
-}
-
-fn evaluate_resident_bytes(
-    bytes: &[u8],
-    policy_store: Option<&PolicySnapshotStore>,
-) -> Result<Vec<u8>, String> {
-    let value = strict_json_value(bytes)?;
-    if value.get("operation").and_then(Value::as_str) == Some("policy_snapshot_push") {
-        let object = value
-            .as_object()
-            .ok_or_else(|| "native_policy_snapshot_push_invalid".to_owned())?;
-        if object
-            .keys()
-            .any(|key| !matches!(key.as_str(), "operation" | "request" | "deadline_budget_ms"))
-        {
-            return Err("native_policy_snapshot_push_invalid".to_owned());
-        }
-    }
-    if value.get("operation").is_none()
-        && value.get("schema").and_then(Value::as_str) != Some(GUARD_HOOK_ENVELOPE_V2_SCHEMA)
-    {
-        oneshot::validate_request_policy_snapshot(&value)?;
-    }
-    let request: ResidentRequestV1 = serde_json::from_value(value)
-        .map_err(|_| "native_resident_request_invalid_json".to_owned())?;
-    match request {
-        ResidentRequestV1::Edge(request) => {
-            let policy_store =
-                policy_store.ok_or_else(|| "native_policy_snapshot_unavailable".to_owned())?;
-            edge::evaluate_envelope_with_store(request, policy_store)
-        }
-        ResidentRequestV1::Operation(ResidentOperationV1::CommandModel(request)) => {
-            oneshot::evaluate_command_model_request(&request)
-        }
-        ResidentRequestV1::Operation(ResidentOperationV1::PreToolUse(request)) => {
-            oneshot::evaluate_pre_tool_request(&request)
-        }
-        ResidentRequestV1::Operation(ResidentOperationV1::PolicySnapshotPush(request)) => {
-            let policy_store =
-                policy_store.ok_or_else(|| "native_policy_snapshot_unavailable".to_owned())?;
-            policy_store.push(&request)
-        }
-        ResidentRequestV1::Operation(ResidentOperationV1::Health(_request)) => {
-            encode_response(&serde_json::json!({
-                "status": "ready",
-                "protocol_version": RESIDENT_PROTOCOL_VERSION,
-            }))
-        }
-        ResidentRequestV1::Operation(ResidentOperationV1::Shutdown(_request)) => {
-            managed_resident::request_shutdown();
-            encode_response(&serde_json::json!({"status": "stopping"}))
-        }
-        ResidentRequestV1::Hook(request) => {
-            // Managed residents always carry a v3 policy store. The legacy
-            // request has no generation-bound snapshot and must never reach a
-            // semantic evaluator in that path.
-            if policy_store.is_some() {
-                Err("native_policy_snapshot_required".to_owned())
-            } else {
-                encode_response(&review_post_tool(&request))
-            }
-        }
-    }
-}
-
-pub(crate) fn encode_response<T: serde::Serialize>(value: &T) -> Result<Vec<u8>, String> {
-    let encoded =
-        serde_json::to_vec(value).map_err(|_| "native_response_encode_failed".to_owned())?;
-    if encoded.len() > MAX_NATIVE_RESPONSE_BYTES {
-        return Err("native_response_too_large".to_owned());
-    }
-    Ok(encoded)
-}
-
-fn error_response(code: &'static str, retryable: bool) -> Vec<u8> {
-    serde_json::to_vec(&serde_json::json!({"error": code, "retryable": retryable})).unwrap_or_else(
-        |_| b"{\"error\":\"native_response_encode_failed\",\"retryable\":false}".to_vec(),
-    )
-}
-
-fn safe_error_response(code: &str, retryable: bool) -> Vec<u8> {
-    if code.starts_with("native_policy_snapshot_") || code.starts_with("snapshot_") {
-        return serde_json::to_vec(&serde_json::json!({
-            "error": code,
-            "retryable": retryable,
-        }))
-        .unwrap_or_else(|_| error_response("native_response_encode_failed", false));
-    }
-    error_response("native_request_invalid_json", retryable)
-}
-
 fn write_json<T: serde::Serialize>(value: &T) -> Result<(), String> {
     serde_json::to_writer(io::stdout().lock(), value)
         .map_err(|_| "native_response_encode_failed".to_owned())?;
     println!();
     Ok(())
-}
-
-fn constant_time_eq(left: &[u8], right: &[u8]) -> bool {
-    if left.len() != right.len() {
-        return false;
-    }
-    let mut difference = 0u8;
-    for (left_byte, right_byte) in left.iter().zip(right) {
-        difference |= left_byte ^ right_byte;
-    }
-    difference == 0
-}
-
-fn hmac_sha256(key: &[u8], label: &[u8], nonce: &[u8]) -> [u8; AUTH_PROOF_BYTES] {
-    const BLOCK_BYTES: usize = 64;
-    let mut key_block = [0u8; BLOCK_BYTES];
-    if key.len() > BLOCK_BYTES {
-        let digest = Sha256::digest(key);
-        key_block[..digest.len()].copy_from_slice(&digest);
-    } else {
-        key_block[..key.len()].copy_from_slice(key);
-    }
-
-    let mut inner_pad = [0x36u8; BLOCK_BYTES];
-    let mut outer_pad = [0x5cu8; BLOCK_BYTES];
-    for index in 0..BLOCK_BYTES {
-        inner_pad[index] ^= key_block[index];
-        outer_pad[index] ^= key_block[index];
-    }
-
-    let mut inner = Sha256::new();
-    inner.update(inner_pad);
-    inner.update(label);
-    inner.update(nonce);
-    let inner_digest = inner.finalize();
-
-    let mut outer = Sha256::new();
-    outer.update(outer_pad);
-    outer.update(inner_digest);
-    let digest = outer.finalize();
-    let mut proof = [0u8; AUTH_PROOF_BYTES];
-    proof.copy_from_slice(&digest);
-    proof
-}
-
-fn authenticate_resident_stream(
-    stream: &mut dyn ResidentStream,
-    token: &[u8; AUTH_TOKEN_BYTES],
-) -> Result<(), String> {
-    stream
-        .set_resident_read_timeout(Some(AUTH_TIMEOUT))
-        .map_err(|_| "native_resident_auth_timeout_failed".to_owned())?;
-    stream
-        .set_resident_write_timeout(Some(AUTH_TIMEOUT))
-        .map_err(|_| "native_resident_auth_timeout_failed".to_owned())?;
-    let _ = stream.configure_low_latency();
-
-    let mut nonce = [0u8; AUTH_NONCE_BYTES];
-    stream
-        .read_exact(&mut nonce)
-        .map_err(|_| "native_resident_auth_nonce_failed".to_owned())?;
-    let server_proof = hmac_sha256(token, SERVER_PROOF_LABEL, &nonce);
-    stream
-        .write_all(&server_proof)
-        .map_err(|_| "native_resident_auth_proof_failed".to_owned())?;
-
-    let mut client_proof = [0u8; AUTH_PROOF_BYTES];
-    stream
-        .read_exact(&mut client_proof)
-        .map_err(|_| "native_resident_auth_client_failed".to_owned())?;
-    let expected = hmac_sha256(token, CLIENT_PROOF_LABEL, &nonce);
-    if !constant_time_eq(&client_proof, &expected) {
-        return Err("native_resident_auth_rejected".into());
-    }
-    Ok(())
-}
-
-fn read_request_header(mut stream: BoxedResidentStream) -> Result<PendingRequest, String> {
-    stream
-        .set_resident_read_timeout(Some(HEADER_TIMEOUT))
-        .map_err(|_| "native_frame_timeout_failed".to_owned())?;
-    let mut header = [0u8; FRAME_HEADER_BYTES];
-    stream
-        .read_exact(&mut header)
-        .map_err(|_| "native_frame_header_failed".to_owned())?;
-    if !constant_time_eq(&header[..4], REQUEST_MAGIC) {
-        return Err("native_frame_version_mismatch".to_owned());
-    }
-    let mut request_id = [0u8; FRAME_REQUEST_ID_BYTES];
-    request_id.copy_from_slice(&header[4..4 + FRAME_REQUEST_ID_BYTES]);
-    let digest_start = 4 + FRAME_REQUEST_ID_BYTES;
-    let mut request_digest = [0u8; FRAME_DIGEST_BYTES];
-    request_digest.copy_from_slice(&header[digest_start..digest_start + FRAME_DIGEST_BYTES]);
-    let length = u32::from_be_bytes(
-        header[FRAME_HEADER_BYTES - 4..]
-            .try_into()
-            .map_err(|_| "native_frame_header_failed".to_owned())?,
-    ) as usize;
-    if length == 0 || length > MAX_NATIVE_REQUEST_BYTES {
-        return Err("native_request_too_large".to_owned());
-    }
-    Ok(PendingRequest {
-        stream,
-        request_id,
-        request_digest,
-        length,
-        payload_prefix: Vec::new(),
-        accepted_at: Instant::now(),
-    })
-}
-
-fn write_bound_response(
-    stream: &mut dyn ResidentStream,
-    request_id: &[u8; FRAME_REQUEST_ID_BYTES],
-    response: &[u8],
-) -> Result<(), String> {
-    if response.is_empty() || response.len() > MAX_NATIVE_RESPONSE_BYTES {
-        return Err("native_response_too_large".to_owned());
-    }
-    stream
-        .set_resident_write_timeout(Some(RESPONSE_TIMEOUT))
-        .map_err(|_| "native_frame_timeout_failed".to_owned())?;
-    let digest = Sha256::digest(response);
-    let mut header = Vec::with_capacity(FRAME_HEADER_BYTES);
-    header.extend_from_slice(RESPONSE_MAGIC);
-    header.extend_from_slice(request_id);
-    header.extend_from_slice(&digest);
-    header.extend_from_slice(&(response.len() as u32).to_be_bytes());
-    stream
-        .write_all(&header)
-        .map_err(|error| hardening::write_error(&error, "native_frame_write_failed"))?;
-    stream
-        .write_all(response)
-        .map_err(|error| hardening::write_error(&error, "native_frame_write_failed"))?;
-    stream
-        .flush()
-        .map_err(|error| hardening::write_error(&error, "native_frame_write_failed"))?;
-    Ok(())
-}
-
-fn write_overload(pending: &mut PendingRequest) {
-    let response = error_response("native_overloaded", true);
-    let _ = write_bound_response(&mut *pending.stream, &pending.request_id, &response);
-}
-
-fn handle_pending_request(mut pending: PendingRequest, policy_store: Option<&PolicySnapshotStore>) {
-    if hardening::request_expired(pending.accepted_at) {
-        let response = error_response("native_request_deadline_exceeded", true);
-        let _ = write_bound_response(&mut *pending.stream, &pending.request_id, &response);
-        return;
-    }
-    let _ = pending
-        .stream
-        .set_resident_read_timeout(Some(PAYLOAD_TIMEOUT));
-    let prefix_length = pending.payload_prefix.len();
-    let mut request = vec![0u8; pending.length];
-    request[..prefix_length].copy_from_slice(&pending.payload_prefix);
-    let response = if pending
-        .stream
-        .read_exact(&mut request[prefix_length..])
-        .is_err()
-    {
-        error_response("native_frame_read_failed", false)
-    } else {
-        let digest = Sha256::digest(&request);
-        if !constant_time_eq(&digest, &pending.request_digest) {
-            error_response("native_request_digest_mismatch", false)
-        } else {
-            match catch_unwind(AssertUnwindSafe(|| {
-                evaluate_resident_bytes(&request, policy_store)
-            })) {
-                Ok(Ok(response)) => response,
-                Ok(Err(reason)) => safe_error_response(&reason, false),
-                Err(_panic) => error_response("native_runtime_panicked", false),
-            }
-        }
-    };
-    let _ = write_bound_response(&mut *pending.stream, &pending.request_id, &response);
-}
-
-fn spawn_workers<T, F>(count: usize, receiver: Receiver<T>, handler: F)
-where
-    T: Send + 'static,
-    F: Fn(T) + Send + Sync + 'static,
-{
-    let receiver = Arc::new(Mutex::new(receiver));
-    let handler = Arc::new(handler);
-    for _ in 0..count {
-        let receiver = Arc::clone(&receiver);
-        let handler = Arc::clone(&handler);
-        thread::spawn(move || loop {
-            let next = match receiver.lock() {
-                Ok(guard) => guard.recv(),
-                Err(_) => return,
-            };
-            match next {
-                Ok(item) => handler(item),
-                Err(_) => return,
-            }
-        });
-    }
-}
-
-fn start_resident_workers(
-    token: Arc<[u8; AUTH_TOKEN_BYTES]>,
-    policy_store: Option<Arc<PolicySnapshotStore>>,
-) -> SyncSender<BoxedResidentStream> {
-    let (evaluation_sender, evaluation_receiver) =
-        sync_channel::<PendingRequest>(EVALUATION_QUEUE_CAPACITY);
-    let evaluation_policy_store = policy_store.clone();
-    spawn_workers(EVALUATION_WORKERS, evaluation_receiver, move |pending| {
-        handle_pending_request(pending, evaluation_policy_store.as_deref());
-    });
-
-    let (authentication_sender, authentication_receiver) =
-        sync_channel::<BoxedResidentStream>(AUTH_QUEUE_CAPACITY);
-    spawn_workers(AUTH_WORKERS, authentication_receiver, move |mut stream| {
-        if authenticate_resident_stream(&mut *stream, &token).is_err() {
-            return;
-        }
-        let mut pending = match read_request_header(stream) {
-            Ok(value) => value,
-            Err(_) => return,
-        };
-        pending
-            .payload_prefix
-            .resize(pending.length.min(AUTHENTICATED_PREFETCH_BYTES), 0);
-        let available = match pending
-            .stream
-            .try_read_available(&mut pending.payload_prefix)
-        {
-            Ok(value) => value,
-            Err(_) => return,
-        };
-        pending.payload_prefix.truncate(available);
-        if available == pending.length {
-            handle_pending_request(pending, policy_store.as_deref());
-            return;
-        }
-        match evaluation_sender.try_send(pending) {
-            Ok(()) => {}
-            Err(TrySendError::Full(returned)) => {
-                pending = returned;
-                write_overload(&mut pending);
-            }
-            Err(TrySendError::Disconnected(_returned)) => {}
-        }
-    });
-    authentication_sender
-}
-
-fn admit_connection(
-    sender: &SyncSender<BoxedResidentStream>,
-    stream: BoxedResidentStream,
-) -> Result<(), String> {
-    match sender.try_send(stream) {
-        Ok(()) => Ok(()),
-        Err(TrySendError::Full(_stream)) => Ok(()),
-        Err(TrySendError::Disconnected(_stream)) => {
-            Err("native_resident_worker_pool_stopped".to_owned())
-        }
-    }
-}
-
-#[cfg(unix)]
-fn resident_parent_liveness() -> Result<Arc<std::sync::atomic::AtomicBool>, String> {
-    let alive = Arc::new(std::sync::atomic::AtomicBool::new(true));
-    let Ok(raw_descriptor) = env::var(PARENT_LIVENESS_FD_ENV) else {
-        return Ok(alive);
-    };
-    let descriptor = raw_descriptor
-        .parse::<u32>()
-        .map_err(|_| "native_parent_liveness_fd_invalid".to_owned())?;
-    let dev_path = format!("/dev/fd/{descriptor}");
-    let proc_path = format!("/proc/self/fd/{descriptor}");
-    let mut pipe = std::fs::File::open(dev_path)
-        .or_else(|_| std::fs::File::open(proc_path))
-        .map_err(|_| "native_parent_liveness_fd_unavailable".to_owned())?;
-    let watcher_state = Arc::clone(&alive);
-    thread::spawn(move || {
-        let mut byte = [0u8; 1];
-        let _ = pipe.read(&mut byte);
-        watcher_state.store(false, std::sync::atomic::Ordering::Release);
-    });
-    Ok(alive)
-}
-
-#[cfg(unix)]
-fn serve(socket_path: &str) -> Result<(), String> {
-    use std::fs;
-    use std::os::unix::fs::{FileTypeExt, PermissionsExt};
-    use std::os::unix::net::UnixListener;
-    use std::path::Path;
-
-    let path = Path::new(socket_path);
-    let parent = path
-        .parent()
-        .ok_or_else(|| "native_socket_parent_missing".to_owned())?;
-    let parent_metadata =
-        fs::symlink_metadata(parent).map_err(|_| "native_socket_parent_stat_failed".to_owned())?;
-    if parent_metadata.file_type().is_symlink()
-        || !parent_metadata.is_dir()
-        || parent_metadata.permissions().mode() & 0o077 != 0
-    {
-        return Err("native_socket_parent_not_private".to_owned());
-    }
-    match fs::symlink_metadata(path) {
-        Ok(metadata) => {
-            if metadata.file_type().is_symlink() || !metadata.file_type().is_socket() {
-                return Err("native_socket_existing_path_rejected".to_owned());
-            }
-            fs::remove_file(path).map_err(|_| "native_socket_cleanup_failed".to_owned())?;
-        }
-        Err(error) if error.kind() == io::ErrorKind::NotFound => {}
-        Err(_) => return Err("native_socket_stat_failed".to_owned()),
-    }
-    let listener = UnixListener::bind(path).map_err(|_| "native_socket_bind_failed".to_owned())?;
-    fs::set_permissions(path, fs::Permissions::from_mode(0o600))
-        .map_err(|_| "native_socket_permissions_failed".to_owned())?;
-    listener
-        .set_nonblocking(true)
-        .map_err(|_| "native_socket_nonblocking_failed".to_owned())?;
-
-    let token = Arc::new(read_resident_auth_token()?);
-    let sender = start_resident_workers(token, None);
-    let parent_alive = resident_parent_liveness()?;
-    let mut consecutive_accept_failures = 0;
-    while parent_alive.load(std::sync::atomic::Ordering::Acquire) {
-        match listener.accept() {
-            Ok((stream, _address)) => {
-                consecutive_accept_failures = 0;
-                if stream.set_nonblocking(false).is_err() {
-                    continue;
-                }
-                admit_connection(&sender, Box::new(stream))?
-            }
-            Err(error)
-                if hardening::classify_io_error(&error) != hardening::IoFailureClass::Other =>
-            {
-                consecutive_accept_failures += 1;
-                thread::sleep(hardening::accept_retry_delay(
-                    consecutive_accept_failures,
-                    &error,
-                ));
-            }
-            Err(_) => return Err("native_socket_accept_failed".to_owned()),
-        }
-    }
-    Ok(())
-}
-
-#[cfg(not(unix))]
-fn serve(_socket_path: &str) -> Result<(), String> {
-    Err("native_unix_socket_not_available".into())
-}
-
-fn serve_loopback(address: &str) -> Result<(), String> {
-    let requested: SocketAddr = address
-        .parse()
-        .map_err(|_| "native_resident_address_invalid".to_owned())?;
-    if requested.ip() != Ipv4Addr::LOCALHOST || requested.port() == 0 {
-        return Err("native_resident_address_not_loopback".into());
-    }
-    let listener = TcpListener::bind(requested)
-        .map_err(|_| "native_resident_loopback_bind_failed".to_owned())?;
-    let local = listener
-        .local_addr()
-        .map_err(|_| "native_resident_loopback_addr_failed".to_owned())?;
-    if local != requested {
-        return Err("native_resident_loopback_addr_changed".into());
-    }
-
-    let token = Arc::new(read_resident_auth_token()?);
-    let sender = start_resident_workers(token, None);
-    let mut consecutive_accept_failures = 0;
-    loop {
-        match listener.accept() {
-            Ok((stream, _address)) => {
-                consecutive_accept_failures = 0;
-                admit_connection(&sender, Box::new(stream))?;
-            }
-            Err(error)
-                if hardening::classify_io_error(&error) != hardening::IoFailureClass::Other =>
-            {
-                consecutive_accept_failures += 1;
-                thread::sleep(hardening::accept_retry_delay(
-                    consecutive_accept_failures,
-                    &error,
-                ));
-            }
-            Err(_) => return Err("native_resident_loopback_accept_failed".to_owned()),
-        }
-    }
-}
-
-fn hex_nibble(value: u8) -> Option<u8> {
-    match value {
-        b'0'..=b'9' => Some(value - b'0'),
-        b'a'..=b'f' => Some(value - b'a' + 10),
-        b'A'..=b'F' => Some(value - b'A' + 10),
-        _ => None,
-    }
-}
-
-fn read_resident_auth_token() -> Result<[u8; AUTH_TOKEN_BYTES], String> {
-    let mut encoded = String::new();
-    io::stdin()
-        .lock()
-        .take((AUTH_TOKEN_BYTES * 2 + 2) as u64)
-        .read_line(&mut encoded)
-        .map_err(|error| hardening::read_error(&error, "native_resident_auth_read_failed"))?;
-    let encoded = encoded.trim();
-    if encoded.len() != AUTH_TOKEN_BYTES * 2 {
-        return Err("native_resident_auth_invalid".into());
-    }
-    let mut token = [0u8; AUTH_TOKEN_BYTES];
-    for (index, pair) in encoded.as_bytes().chunks_exact(2).enumerate() {
-        let high = hex_nibble(pair[0]).ok_or_else(|| "native_resident_auth_invalid".to_owned())?;
-        let low = hex_nibble(pair[1]).ok_or_else(|| "native_resident_auth_invalid".to_owned())?;
-        token[index] = (high << 4) | low;
-    }
-    Ok(token)
-}
-
-fn resident_stdin_liveness() -> Arc<std::sync::atomic::AtomicBool> {
-    let alive = Arc::new(std::sync::atomic::AtomicBool::new(true));
-    let watcher_state = Arc::clone(&alive);
-    thread::spawn(move || {
-        let mut byte = [0u8; 1];
-        let _ = io::stdin().read(&mut byte);
-        watcher_state.store(false, std::sync::atomic::Ordering::Release);
-    });
-    alive
 }
 
 fn write_bytes_response(response: &[u8]) -> Result<(), String> {
